Refactor : CommentController Refactoring -CommentController 댓글 관련 리팩토링 (jwt 토큰 추가) Resolves : Jira AI02-51

+++ findFounder/src/main/java/com/example/demo/service/CommentService.java
@@ -3,6 +3,7 @@ package com.example.demo.service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -47,8 +48,9 @@ public class CommentService {
 	
 	// Create Comment
 	@Transactional
-	public CommonRes createComment(CommentCreateReq req) {
-		Customer customer = customerRepository.findById(req.getCusCode()).orElseThrow(() -> new UsernameNotFoundException("해당하는 유저가 없습니다."));
+	public CommonRes createComment(int cuscode, CommentCreateReq req) {
+//		Customer customer = customerRepository.findById(req.getCusCode()).orElseThrow(() -> new UsernameNotFoundException("해당하는 유저가 없습니다."));
+		Customer customer = customerRepository.findById(cuscode).orElseThrow(() -> new UsernameNotFoundException("해당하는 사용자가 없습니다."));
 		Post post = postService.getPost(req.getPostId());
 		Comment comment = Comment.builder()
 							.commentContent(req.getCommentContent())
@@ -56,19 +58,37 @@ public class CommentService {
 							.customer(customer)
 							.build();
 		commentRepository.saveAndFlush(comment);
-		CommonRes commonRes = CommonRes.builder().code(200).msg("댓글 등록 완료").build();
+		CommonRes commonRes = CommonRes.builder().code(200).msg("댓글 등록 완료되었습니다.").build();
 		return commonRes;
 	}
 	
 	// Update Comment
 	@Transactional
-	public CommonRes updateComment(CommentUpdateReq req) {
-		Comment comment = findComment(req.getCommentId());
-		comment.setCommentContent(req.getCommentContent());
+	public CommonRes updateComment(int cuscode, CommentUpdateReq req) {
+		Optional<Customer> customerOptional = customerRepository.findById(cuscode);
 		
-		commentRepository.save(comment);
+		CommonRes commonRes; 
+		if (customerOptional.isPresent()) {
+
+			Comment comment = findComment(req.getCommentId());
+
+			if (comment.getCustomer().getCusCode() == cuscode) {
+				comment.setCommentContent(req.getCommentContent());
 		
-		CommonRes commonRes = CommonRes.builder().code(200).msg("댓글 수정 완료").build();
+				commentRepository.save(comment);
+		
+				commonRes = CommonRes.builder().code(200).msg("댓글 수정 완료되었습니다.").build();
+				
+			}
+			else {
+				commonRes = CommonRes.builder().code(300).msg("본인이 아니기 때문에 댓글 수정이 불가능합니다.").build();
+			}
+			
+		}
+		else {
+			commonRes = CommonRes.builder().code(400).msg("회원 정보가 존재하지 않습니다.").build();
+		}
+		
 		return commonRes;
 	}
 	
@@ -83,10 +103,27 @@ public class CommentService {
 	
 	// delete Comment
 	@Transactional
-	public CommonRes deleteComment(int commentId) {
-		commentRepository.deleteById(commentId);
-		CommonRes commonRes = CommonRes.builder().code(200).msg("댓글 삭제 완료").build();
+	public CommonRes deleteComment(int cuscode, int commentId) {
+		Optional<Customer> customerOptional = customerRepository.findById(cuscode);
+		
+		CommonRes commonRes; 
+		if (customerOptional.isPresent()) {
+			Comment comment = findComment(commentId);
+			if (comment.getCustomer().getCusCode() == cuscode) {
+				commentRepository.deleteById(commentId);
+				commonRes = CommonRes.builder().code(200).msg("댓글 삭제 완료되었습니다.").build();
+			}
+			else {
+			commonRes = CommonRes.builder().code(300).msg("본인이 아니기 때문에 댓글 삭제가 불가능합니다.").build();
+			}
+		} 
+		else {
+		commonRes = CommonRes.builder().code(400).msg("회원 정보가 존재하지 않습니다.").build();
+		}
 		return commonRes;
+			
+		}
+			
+		
 	}
 	
-}
